requirements.txt
absl-py==2.0.0
accelerate==1.5.2
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==4.0.2
attrs==23.1.0
av==14.2.0
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.6
colorama==0.4.6
contourpy==1.1.1
cryptography==44.0.2
cycler==0.11.0
datasets==2.19.1
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
dpcpp-cpp-rt==2025.0.5
einops==0.6.1
emoji==0.6.0
exceptiongroup==1.2.0
executing==2.2.0
fastlid==0.1.9
fasttext==0.9.2
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.0
google-auth-oauthlib==1.0.0
googletrans==4.0.2
gpustat==1.1
grpcio==1.58.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.3
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
intel-cmplr-lib-rt==2025.0.5
intel-cmplr-lib-ur==2025.0.5
intel-cmplr-lic-rt==2025.0.5
intel-opencl-rt==2025.0.5
intel-openmp==2025.0.5
intel-sycl-rt==2025.0.5
ipython==8.18.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
langid==1.1.6
lingua-language-detector==2.0.2
logzero==1.7.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.7
mkl==2025.0.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
natsort==8.4.0
networkx==3.2.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.535.77
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.67.0
openprompt==1.0.1
packaging==23.1
pandas==2.2.2
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.4.0
prompt_toolkit==3.0.50
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qwen-vl-utils==0.0.10
regex==2023.8.8
requests==2.32.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
rouge==1.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.96
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
sympy==1.13.1
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
textblob==0.19.0
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.0
traitlets==5.14.3
transformers==4.49.0
transformers-stream-generator==0.0.4
triton==3.2.0
tweet-preprocessor==0.6.0
typing_extensions==4.12.2
tzdata==2024.1
umf==0.9.1
urllib3==1.26.20
wandb==0.15.12
wcwidth==0.2.6
Werkzeug==2.3.7
wordninja==2.0.0
xxhash==3.4.1
yacs==0.1.8
yaml-kit==0.0.1
yarl==1.9.2
zipp==3.17.0
